Finished first version of airport size calculations
Also added A380 support column to airports
Also added size 6 airports based on top50 busiest in the world
