+++ OpenSky.API/Datasets/a380airports.txt
@@ -1,0 +1,393 @@
+﻿OISS
+OIZH
+KABQ
+DNAA
+LEMG
+SGES
+KAMA
+DIAP
+HAAB
+YPAD
+NZAA
+UAAA
+EHAM
+OJAI
+ESSA
+KATL
+OMAA
+OBBI
+LEBL
+EGBB
+VTBS
+VOBL
+YBBN
+SKBO
+VABB
+KBOS
+EBBR
+HECA
+ZGGG
+LFPG
+WIII
+VCBI
+GMMN
+EKCH
+FACT
+ZUUU
+MMUN
+VIDP
+KDEN
+KDFW
+UUDD
+OTHH
+FALE
+EDDL
+OMDW
+OMDB
+ORER
+EGNX
+SAEZ
+LIRF
+EDDF
+SBGL
+EGPF
+SBGR
+UBBB
+EDDH
+VVNB
+ZSHC
+VHHH
+RJTT
+PHNL
+VOHS
+KIAD
+KIAH
+RKSI
+OPRN
+LTBA
+OEJN
+KJFK
+FAOR
+UHHH
+RJBB
+ZPPP
+WIMM
+WMKK
+OKBK
+KLAS
+KLAX
+EGKK
+EGLL
+DNMM
+LEMD
+EGCC
+KMCO
+OOMS
+OEMA
+YMML
+MMMX
+KMIA
+RPLL
+FIMP
+EDDM
+LIMC
+DTNH
+RJGG
+RJAA
+KORD
+ENGM
+ZBAA
+LEPA
+LKPR
+MDPC
+ZSPD
+OERK
+SCEL
+KSFO
+VVTS
+ZSSS
+WSSS
+OOSA
+UUEE
+YSSY
+ZGSZ
+UTTT
+LFBO
+LLBG
+ZSJN
+RCTP
+ZBTJ
+SBKP
+LOWW
+UUWW
+UHWW
+ZLXY
+CYUL
+CYVR
+CYYZ
+LSZH
+ZGSD
+OMAL
+UATT
+PANC
+NSFA
+UTAA
+YBAS
+LGAV
+YMAV
+KBDL
+LFRB
+KBGR
+KBHM
+LFBD
+EDDW
+EGGD
+WIDD
+LHBP
+EDDK
+NZCH
+LFLX
+RKPC
+VTCC
+RJCC
+KCVG
+EGFF
+GOOY
+VTBD
+EDDC
+YPDN
+LTAC
+TFFF
+EDSB
+GCFV
+RKSS
+LSGG
+EDDV
+EFHK
+ZBHH
+OIIE
+PHTO
+WMKJ
+UKBB
+BIKF
+OPKC
+PHKO
+LFBT
+EDDP
+LOWL
+GCLP
+LFLL
+SKRG
+KMKE
+LMML
+KMSP
+NFFN
+LFMN
+NWWW
+LFRS
+KOAK
+KONT
+LFPO
+UNNT
+LFMP
+VIPT
+KPHL
+NSTU
+NTAA
+TFFR
+WSAP
+YBRK
+FMEE
+UKFF
+ZBSJ
+KSLC
+EINN
+HESH
+EGSS
+EDDS
+LEZL
+USSS
+KSWF
+EDDB
+LTCG
+VTBU
+EPWA
+LFOK
+EDHI
+CYEG
+CYFB
+CYHZ
+CYOW
+CYQX
+DGAA
+SGAS
+TNCA
+SKBQ
+UNBB
+TXKF
+OLBA
+KBFI
+EGAA
+FEFF
+TBPB
+GABS
+LIPE
+LFSB
+KBWI
+SVMI
+VECC
+PACD
+RPVM
+ZHCC
+ZUCK
+KCLE
+VOCI
+DBBB
+KCOS
+RPLC
+TNCC
+VGHS
+VVDN
+KDAL
+HTDA
+KDAY
+FKKD
+ZYTL
+OEDF
+KDTW
+UHMA
+HUEN
+KELP
+UDYZ
+PAFA
+FZAA
+ZSFZ
+UCFM
+FBSK
+KGEG
+KGTF
+UATG
+SEGU
+ZJHK
+EDFH
+RJOA
+RJCH
+VTSP
+WIHH
+ZYHB
+FVHA
+VCRI
+KIAG
+KIND
+SCDA
+KJAX
+HTKJ
+DNKN
+VTSG
+WBGG
+UNEE
+UAKK
+ZWSH
+RCKH
+UNKL
+UTAK
+YPTN
+EPKT
+UBBG
+LCLK
+YPLM
+DXXX
+KLGB
+EBLG
+OPLA
+SPIM
+FWKI
+SELT
+FLKK
+ELLX
+ENSB
+VOMM
+SVMC
+HKMO
+KMCI
+VYMD
+SAME
+KMEM
+KMHR
+FQMA
+UMMS
+OPMT
+SUMU
+KMWH
+HKJK
+EGNT
+ZSNB
+ZSNJ
+FKYS
+UOOO
+KOKC
+EBOS
+KPBI
+KPDX
+KPHX
+EGPK
+SPSO
+KPIT
+SVMG
+TTPP
+KPSM
+MPTO
+RKPK
+GMME
+EVRA
+KRSW
+OYSN
+MSLP
+KSAN
+KSAT
+KSAV
+LTFJ
+KSBD
+KSDF
+MDSD
+KSEA
+ZYTX
+OMSJ
+GVAC
+TJSJ
+UTSS
+KSTL
+WARR
+ENZV
+ZSQD
+LPLA
+GCTS
+OETF
+KTPA
+LIMF
+RCSS
+SEQM
+ZWWW
+LIPO
+LIPZ
+LEVC
+FYWH
+ZHHH
+ZSAM
+ZSYT
+CYQM
+CYWG
+CYYC
+CYYR
+LEZG
+HTZA
+OIMM
+UEEE
+ESGG
+ESMS
+UWLW
+UTSA
+ZLHX
+KAUS
+EGMH
+YPPH
