+++ OpenSky.API/Datasets/top50airports.txt
@@ -1,0 +1,50 @@
+KATL
+ZBAA
+KLAX
+OMDB
+RJTT
+KORD
+EGLL
+ZSPD
+LFPG
+KDFW
+ZGGG
+EHAM
+VHHH
+RKSI
+EDDF
+KDEN
+VIDP
+WSSS
+VTBS
+KJFK
+WMKK
+LEMD
+KSFO
+ZUUU
+WIII
+ZGSZ
+LEBL
+LTFM
+KSEA
+KLAS
+KMCO
+CYYZ
+MMMX
+KCLT
+UUEE
+RCTP
+ZPPP
+EDDM
+RPLL
+ZLXY
+VABB
+EGKK
+KEWR
+KPHX
+KMIA
+ZSSS
+KIAH
+ZUCK
+YSSY
+RJAA
